Separate files for element flows and sink. Added tag path flow for elements
